Squashed commit of the following:
commit febf53fc7ef894a781d60221e7bca06a99c9a3de
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 04:55:39 2024 +0200

    remove unused userId from body of some requests

commit ab0624aca8f1f9441788e20567a51d95577c084c
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 04:35:46 2024 +0200

    Same for editing a comment

commit cabf1d655dbfc1405da920d202c481ca7be0e426
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:44:07 2024 +0200

    Authinticate in delete comment too

commit c40115c089b9d729fb4f3714a0fcc05d606bbc90
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:29:01 2024 +0200

    Same for adding comments

commit 74974060ec2cf1b83b32e316a8e75629f6a55932
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:27:13 2024 +0200

    Implement JWT in fetching comments

+++ test_server/routes/comments.js
@@ -11,12 +11,14 @@ const {
   repliesList,
 } = require('../mockData');
 const db = require('../connect');
-const { getUserData } = require('../helperFunctions');
+const { getUserData, isCourseAdmin } = require('../helperFunctions');
+const { verifyToken } = require('../middlewares/authMiddlewares');
+
 
 const router = express.Router();
 
 // Get all comments for an announcement
-router.get('/announcements/:id/comments', (req, res) => {
+router.get('/announcements/:id/comments', verifyToken, (req, res) => {
   const announcementId = req.params.id;
   const announcement = db.prepare('SELECT 1 FROM announcements WHERE id = ?').get(announcementId);
 
@@ -47,11 +49,12 @@ router.get('/announcements/:id/comments', (req, res) => {
 });
 
 // Create a comment for an announcement
-router.post('/announcements/:id/comments', (req, res) => {
+router.post('/announcements/:id/comments', verifyToken, (req, res) => {
   const announcementId = req.params.id;
-  const { userId, comment: body } = req.body;
+  const {comment: body } = req.body;
+  const userId = req.userId;
 
-  if (!userId || !body) {
+  if (!body) {
     return res.status(400).send({ message: 'Missing required fields' });
   }
   
@@ -81,15 +84,20 @@ router.post('/announcements/:id/comments', (req, res) => {
 });
 
 // Edit an announcement
-router.put('/comments/:id', (req, res) => {
+router.put('/comments/:id', verifyToken, (req, res) => {
   const { id } = req.params;
   const { body } = req.body;
+  const userId = req.userId;
 
   try {
-    const comment = db.prepare('SELECT 1 FROM comments WHERE id = ?').get(id);
+    const comment = db.prepare('SELECT userId FROM comments WHERE id = ?').get(id);
 
     if (!comment) {
       return res.status(404).send({ message: 'Comment not found' });
+    }
+
+    if (comment.userId !== userId) {
+      return res.status(403).send({ message: 'User is not authorized to edit this comment' });
     }
 
     db.prepare('UPDATE comments SET body = ? WHERE id = ?').run(body, id);
@@ -106,14 +114,18 @@ router.put('/comments/:id', (req, res) => {
 });
 
 // Delete an announcement;
-router.delete('/comments/:commentId', (req, res) => {
+router.delete('/comments/:commentId', verifyToken, (req, res) => {
   const { commentId } = req.params;
+  const userId = req.userId;
 
   try {
-    const comment = db.prepare('SELECT * FROM comments WHERE id = ?').get(commentId);
-
+    const comment = db.prepare('SELECT userId, announcementId FROM comments WHERE id = ?').get(commentId);
     if (!comment) {
       return res.status(404).send({ message: 'Comment not found' });
+    }
+
+    if (comment.userId !== userId && !isCourseAdmin(userId, comment.announcementId)) {
+      return res.status(403).send({ message: 'User is not a course admin' });
     }
 
     db.prepare('DELETE FROM comments WHERE id = ?').run(commentId);
